src/requirements.txt
pydantic==2.4.2
pydantic[email]
supabase
pydantic-settings
python-dotenv
fastapi
uvicorn
PyJWT